+++ tests/integration/test_migrations.py
@@ -11,8 +11,9 @@ from minumtium_postgres.migrations import (has_version_table,
                                            MigrationVersion,
                                            run_migrations,
                                            apply_migrations,
-                                           update_database_version,
-                                           MIGRATION_TABLE_NAME)
+                                           update_database_version)
+
+from minumtium_postgres.migrations.versions.version_0 import Version0
 
 
 def test_has_version_table(database_with_version_table):
@@ -28,12 +29,12 @@ def test_get_database_version(database_with_version_table):
 
 
 def test_run_migrations(mock_migration, mock_table_name, database, schema):
-    run_migrations([mock_migration], database, schema)
+    run_migrations([mock_migration], database, "", schema)
     assert inspect(database).has_table(mock_table_name)
 
 
 def test_apply_any_migrations(some_migrations, mock_table_name, another_mock_table_name, database, schema):
-    run_migrations(some_migrations, database, schema)
+    run_migrations(some_migrations, database, "", schema)
     assert inspect(database).has_table(mock_table_name)
     assert inspect(database).has_table(another_mock_table_name)
 
@@ -75,7 +76,7 @@ def some_migrations(another_mock_table_name, mock_migration) -> List[MigrationVe
         def get_version(self) -> int:
             return 1
 
-        def do(self, engine, schema) -> None:
+        def do(self, engine, schema, **kwargs) -> None:
             meta = MetaData(schema=schema)
 
             table = Table(
@@ -84,7 +85,7 @@ def some_migrations(another_mock_table_name, mock_migration) -> List[MigrationVe
             )
             meta.create_all(engine)
 
-        def undo(self, engine) -> None:
+        def undo(self, engine, **kwargs) -> None:
             pass
 
     return [mock_migration, AnotherMockMigration()]
@@ -96,7 +97,7 @@ def mock_migration(mock_table_name) -> MigrationVersion:
         def get_version(self) -> int:
             return 0
 
-        def do(self, engine, schema) -> None:
+        def do(self, engine, schema, **kwargs) -> None:
             meta = MetaData(schema=schema)
 
             table = Table(
@@ -105,7 +106,7 @@ def mock_migration(mock_table_name) -> MigrationVersion:
             )
             meta.create_all(engine)
 
-        def undo(self, engine) -> None:
+        def undo(self, engine, **kwargs) -> None:
             pass
 
     return MockMigration()
@@ -120,13 +121,13 @@ def schema():
 def database_with_version_table(database):
     def create_version_table(database):
         meta = MetaData()
-        Table(MIGRATION_TABLE_NAME, meta, Column('version', Integer))
+        Table(Version0.migration_table_name(), meta, Column('version', Integer))
         meta.create_all(database)
 
     def insert_version(database, version):
         with database.connect() as connection:
             with connection.begin():
-                statement = text(f"INSERT INTO {MIGRATION_TABLE_NAME} VALUES(:version)")
+                statement = text(f"INSERT INTO {Version0.migration_table_name()} VALUES(:version)")
                 connection.execute(statement, {'version': version})
 
     create_version_table(database)