+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -6,10 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2022-06-5
+
+### Changed
+
+- Changed adapter to get a prefix name.
+- Changed migrations to get a table prefix name.
+
 ## [1.0.1] - 2022-05-22
+
 ### Fix
+
 - Adding missing cast in id field that was breaking the delete method.
 
 ## [1.0.0] - 2022-04-24
+
 ### Added
+
 - Initial version of database adapter for postgres database.